+++ libs/oef-core/src/conversations/SearchRemoveTask.cpp
@@ -57,7 +57,7 @@ SearchRemoveTask::StateResult SearchRemoveTask::HandleResponse(void)
       FETCH_LOG_WARN(LOGGING_NAME, "  ", n);
     }
     std::shared_ptr<OUT_PROTO> answer = std::make_shared<OUT_PROTO>();
-    answer->set_answer_id(msg_id_);
+    answer->set_answer_id(static_cast<int32_t>(msg_id_));
     auto error = answer->mutable_oef_error();
     error->set_operation(fetch::oef::pb::Server_AgentMessage_OEFError::UNREGISTER_SERVICE);
     FETCH_LOG_WARN(LOGGING_NAME, "Sending error ", error, " to ", agent_uri_);
